Refactor ZoomingHandler to CameraZoomHandler Jetty
