+++ backend/src/main/java/com/benine/backend/http/FocussingHandler.java
@@ -4,6 +4,7 @@ import com.benine.backend.LogEvent;
 import com.benine.backend.camera.Camera;
 import com.benine.backend.camera.CameraConnectionException;
 import com.benine.backend.camera.FocussingCamera;
+import com.benine.backend.http.jetty.MalformedURIException;
 import com.sun.net.httpserver.HttpExchange;
 
 import java.io.IOException;
