+++ backend/src/main/java/com/benine/backend/http/jetty/MalformedURIException.java
@@ -1,4 +1,4 @@
-package com.benine.backend.http;
+package com.benine.backend.http.jetty;
 
 /**
  * Created on 4-5-16.
